% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bom.R
\name{bom_breakdown}
\alias{bom_breakdown}
\title{Title bom展开函数}
\usage{
bom_breakdown(
  file_name = "data-raw/需求说明文档LD.xlsx",
  sheet_data = "2023新模板表",
  sheet_BOM = "2023BOM表"
)
}
\arguments{
\item{file_name}{文件名}

\item{sheet_data}{data页签}

\item{sheet_BOM}{BOM页签}
}
\value{
返回值
}
\description{
Title bom展开函数
}
\examples{
bom_breakdown()
}
