% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_dealing.R
\name{mo_combine}
\alias{mo_combine}
\title{针对生产订单按工事番号进行合并}
\usage{
mo_combine(file_name = "data-raw/原始数据v0.4.xlsx", key_word = "轿顶站")
}
\arguments{
\item{file_name}{文件名}

\item{key_word}{关键词}
}
\value{
返回值
}
\description{
针对生产订单按工事番号进行合并
}
\examples{
mo_combine()
}
