% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formattor_23to22.R
\name{mo_formatter}
\alias{mo_formatter}
\title{生产订单2023版转2022版}
\usage{
mo_formatter(
  file_name = "data-raw/新三菱订单模板2023转2022.xlsx",
  sheet_name = "2023新模板表"
)
}
\arguments{
\item{file_name}{文件名}

\item{sheet_name}{页签名}
}
\value{
返回处理后数据集
}
\description{
生产订单2023版转2022版
}
\examples{
mo_formatter()
}
